Use de Bruijn levels as names in ocaml translation.
This results in a hefty simplification of the code and quite a few
lines shaved off. This is also a more satisfying solution to the
repeated variable name problem. Only downside is that the produced
code becomes less readable because the original names for all locally
bound variables are lost.