Isolate a few types of Goptions in a pure .mli, solving a dep issue with ocamlbuild
A few types of Goptions are mentioned in toplevel/interface.mli.
Since the latter is a pure .mli, this shouldn't trigger a dependency
with respect to goptions.ml, but apparently ocamlbuild isn't clever
enough to notice this, and hence wants to link a pile of useless stuff
with coqide. I'll discuss with Pierre-Marie about the best long-term
way to avoid this mess, but in the meantime I split the concerned types
of Goptions in a separate Goptionstyp.mli.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/coq/trunk@14841 85f007b7-540e-0410-9357-904b9bb8a0f7